object unit tests
