+++ internal/pkg/object/object_test.go
@@ -1,6 +1,18 @@
 package object
 
-/*func TestNameObjectCompareTo(t *testing.T) {
+import (
+	"fmt"
+	"math/rand"
+	"reflect"
+	"sort"
+	"testing"
+
+	"github.com/netsec-ethz/rains/internal/pkg/algorithmTypes"
+	"github.com/netsec-ethz/rains/internal/pkg/keys"
+	"golang.org/x/crypto/ed25519"
+)
+
+func TestNameObjectCompareTo(t *testing.T) {
 	nos := sortedNameObjects(9)
 	var shuffled []Name
 	for _, no := range nos {
@@ -18,87 +30,13 @@ package object
 	}
 }
 
-func TestSignatureAlgorithmTypeString(t *testing.T) {
-	var tests = []struct {
-		input SignatureAlgorithmType
-		want  string
-	}{
-		{SignatureAlgorithmType(0), "0"},
-		{Ed25519, "1"},
-		{Ed448, "2"},
-		{Ecdsa256, "3"},
-		{Ecdsa384, "4"},
-	}
-	for i, test := range tests {
-		if test.input.String() != test.want {
-			t.Errorf("%d: Wrong Signature algorithm String value. expected=%v, actual=%v", i, test.want, test.input.String())
-		}
-	}
-}
-
-func TestPublicKeyIDString(t *testing.T) {
-	var tests = []struct {
-		input PublicKeyID
-		want  string
-	}{
-		{PublicKeyID{}, "AT=0 KS=0 KP=0"},
-		{PublicKeyID{Algorithm: Ed25519, KeySpace: RainsKeySpace, KeyPhase: 2}, "AT=1 KS=0 KP=2"},
-	}
-	for i, test := range tests {
-		if test.input.String() != test.want {
-			t.Errorf("%d: Wrong public key id String value. expected=%v, actual=%v", i, test.want, test.input.String())
-		}
-	}
-}
-
-func TestPublicKeyString(t *testing.T) {
-	var tests = []struct {
-		input PublicKey
-		want  string
-	}{
-		{PublicKey{}, "{AT=0 KS=0 KP=0 VS=0 VU=0 data=}"},
-		{
-			PublicKey{
-				PublicKeyID: PublicKeyID{
-					Algorithm: Ed25519,
-					KeySpace:  RainsKeySpace,
-					KeyPhase:  1,
-				},
-				ValidSince: 1,
-				ValidUntil: 2,
-				Key:        ed25519.PublicKey([]byte("PublicKeyData"))},
-			"{AT=1 KS=0 KP=1 VS=1 VU=2 data=5075626c69634b657944617461}",
-		},
-	}
-	for i, test := range tests {
-		if test.input.String() != test.want {
-			t.Errorf("%d: Wrong public key String value. expected=%v, actual=%v", i, test.want, test.input.String())
-		}
-	}
-}
-
-func TestCertTypeString(t *testing.T) {
-	var tests = []struct {
-		input Certificate
-		want  string
-	}{
-		{Certificate{}, "{0 0 0 }"},
-		{Certificate{Type: PTTLS, Usage: CUEndEntity, HashAlgo: Sha256, Data: []byte("testCert")}, "{1 3 1 7465737443657274}"},
-	}
-	for i, test := range tests {
-		if test.input.String() != test.want {
-			t.Errorf("%d: Wrong cert type String value. expected=%v, actual=%v", i, test.want, test.input.String())
-		}
-	}
-}
-
 func TestPublicKeyIDHash(t *testing.T) {
 	var tests = []struct {
-		input PublicKeyID
+		input keys.PublicKeyID
 		want  string
 	}{
-		{PublicKeyID{}, "0,0,0"},
-		{PublicKeyID{Algorithm: Ed25519, KeySpace: RainsKeySpace, KeyPhase: 2}, "1,0,2"},
+		{keys.PublicKeyID{}, "0,0,0"},
+		{keys.PublicKeyID{Algorithm: algorithmTypes.Ed25519, KeySpace: keys.RainsKeySpace, KeyPhase: 2}, "1,0,2"},
 	}
 	for i, test := range tests {
 		if test.input.Hash() != test.want {
@@ -109,15 +47,15 @@ func TestPublicKeyIDHash(t *testing.T) {
 
 func TestPublicKeyHash(t *testing.T) {
 	var tests = []struct {
-		input PublicKey
+		input keys.PublicKey
 		want  string
 	}{
-		{PublicKey{}, "0,0,0,0,0,"},
+		{keys.PublicKey{}, "0,0,0,0,0,"},
 		{
-			PublicKey{
-				PublicKeyID: PublicKeyID{
-					Algorithm: Ed25519,
-					KeySpace:  RainsKeySpace,
+			keys.PublicKey{
+				PublicKeyID: keys.PublicKeyID{
+					Algorithm: algorithmTypes.Ed25519,
+					KeySpace:  keys.RainsKeySpace,
 					KeyPhase:  1,
 				},
 				ValidSince: 1,
@@ -135,7 +73,7 @@ func TestPublicKeyHash(t *testing.T) {
 
 func TestPublicKeyCompareTo(t *testing.T) {
 	pks := sortedPublicKeys(9)
-	var shuffled []PublicKey
+	var shuffled []keys.PublicKey
 	for _, pk := range pks {
 		shuffled = append(shuffled, pk)
 	}
@@ -158,11 +96,11 @@ func TestPublicKeyCompareTo(t *testing.T) {
 		t.Error("Error case was not hit")
 	}
 
-	pk1.KeySpace = KeySpaceID(1)
+	pk1.KeySpace = keys.KeySpaceID(1)
 	if pk1.CompareTo(pks[0]) != 1 {
 		t.Error("key space comparison")
 	}
-	pk1.KeySpace = KeySpaceID(-1)
+	pk1.KeySpace = keys.KeySpaceID(-1)
 	if pk1.CompareTo(pks[0]) != -1 {
 		t.Error("key space comparison")
 	}
@@ -205,7 +143,7 @@ func TestServiceInfoCompareTo(t *testing.T) {
 }
 
 func TestObjectCompareTo(t *testing.T) {
-	objs := sortedObjects(13)
+	objs := SortedObjects(13)
 	var shuffled []Object
 	for _, obj := range objs {
 		shuffled = append(shuffled, obj)
@@ -226,7 +164,7 @@ func TestObjectCompareTo(t *testing.T) {
 	if obj1.CompareTo(objs[0]) != 0 {
 		t.Error("Error case was not hit")
 	}
-	obj1.Value = PublicKey{}
+	obj1.Value = keys.PublicKey{}
 	if obj1.CompareTo(objs[0]) != 0 {
 		t.Error("Error case was not hit")
 	}
@@ -251,7 +189,7 @@ func TestObjectCompareTo(t *testing.T) {
 	}
 }
 func TestObjectString(t *testing.T) {
-	obj := GetAllValidObjects()
+	obj := AllObjects()
 	var tests = []struct {
 		input Object
 		want  string
@@ -261,15 +199,15 @@ func TestObjectString(t *testing.T) {
 		{obj[1], "OT:2 OV:2001:db8::"},
 		{obj[2], "OT:3 OV:192.0.2.0"},
 		{obj[3], "OT:4 OV:example.com"},
-		{obj[4], fmt.Sprintf("OT:5 OV:%s", obj[4].Value.(PublicKey).String())},
+		{obj[4], fmt.Sprintf("OT:5 OV:%s", obj[4].Value.(keys.PublicKey).String())},
 		{obj[5], "OT:6 OV:Would be an expression"},
 		{obj[6], fmt.Sprintf("OT:7 OV:%s", obj[6].Value.(Certificate).String())},
 		{obj[7], "OT:8 OV:{srvName 49830 1}"},
 		{obj[8], "OT:9 OV:Registrar information"},
 		{obj[9], "OT:10 OV:Registrant information"},
-		{obj[10], fmt.Sprintf("OT:11 OV:%s", obj[10].Value.(PublicKey).String())},
-		{obj[11], fmt.Sprintf("OT:12 OV:%s", obj[11].Value.(PublicKey).String())},
-		{obj[12], fmt.Sprintf("OT:13 OV:%s", obj[12].Value.(PublicKey).String())},
+		{obj[10], fmt.Sprintf("OT:11 OV:%s", obj[10].Value.(keys.PublicKey).String())},
+		{obj[11], fmt.Sprintf("OT:12 OV:%s", obj[11].Value.(keys.PublicKey).String())},
+		{obj[12], fmt.Sprintf("OT:13 OV:%s", obj[12].Value.(keys.PublicKey).String())},
 	}
 	for i, test := range tests {
 		if test.input.String() != test.want {
@@ -292,4 +230,4 @@ func TestObjectSort(t *testing.T) {
 	//error case
 	obj = Object{Type: OTExtraKey, Value: ""}
 	obj.Sort()
-}*/
+}
